Isolating update-related tests
