+++ src/test/java/org/smof/BasicSmofTest.java
@@ -46,7 +46,6 @@ import org.smof.dataModel.Owner;
 import org.smof.dataModel.TypeGuitar;
 import org.smof.element.AbstractElement;
 import org.smof.element.Element;
-import org.smof.exception.SmofException;
 import org.smof.gridfs.SmofGridStreamManager;
 import org.smof.index.InternalIndex;
 
@@ -175,7 +174,7 @@ public class BasicSmofTest {
 	}
 	
 	@Test
-	public void testUpdateReplace() {
+	public void testReplace() {
 		final Brand brand = Brand.create("Gibson", new Location("Nashville", "USA"), Collections.singletonList(OWNER_1));
 		smof.insert(brand);
 		brand.setCapital(1000);
@@ -183,79 +182,6 @@ public class BasicSmofTest {
 		assertEquals(brand, smof.find(Brand.class).byElement(brand));
 	}
 	
-	@Test
-	public void testUpdateIncrease() {
-		final Brand brand = Brand.create("Gibson", new Location("Nashville", "USA"), Collections.singletonList(OWNER_1));
-		final long inc = 75L;
-		smof.insert(brand);
-		smof.update(Brand.class)
-		.increase(Brand.CAPITAL, inc)
-		.where()
-		.fieldEq(Brand.NAME, "Gibson")
-		.execute();
-		brand.increaseCapital(inc);
-		final Brand actual = smof.find(Brand.class).byElement(brand);
-		assertEquals(brand, actual);
-	}
-
-	@Test(expected = SmofException.class)
-	public void testUpdateUnknownField() {
-		final Brand brand = Brand.create("Gibson", new Location("Nashville", "USA"), Collections.singletonList(OWNER_1));
-		smof.insert(brand);
-		smof.update(Brand.class)
-		.where()
-		.fieldEq("unknown", "Gibson")
-		.execute();
-	}
-
-	@Test
-	public void testUpdateMultiply() {
-		final Brand brand = Brand.create("Gibson", new Location("Nashville", "USA"), Collections.singletonList(OWNER_1));
-		brand.setCapital(4L);
-		final long mul = 2L;
-		smof.insert(brand);
-		smof.update(Brand.class)
-		.multiply(Brand.CAPITAL, mul)
-		.where()
-		.fieldEq(Brand.NAME, "Gibson")
-		.execute();
-		brand.multiplyCapital(mul);
-		final Brand actual = smof.find(Brand.class).byElement(brand);
-		assertEquals(brand, actual);
-	}
-
-	@Test
-	public void testUpdateSet() {
-		final Brand brand = Brand.create("Gibson", new Location("Nashville", "USA"), Collections.singletonList(OWNER_1));
-		smof.insert(brand);
-		Location newLocation = new Location("New York", "USA");
-		smof.update(Brand.class)
-		.set(Brand.LOCATION, newLocation)
-		.where()
-		.fieldEq(Brand.NAME, "Gibson")
-		.execute();
-		final Brand expected = Brand.create("Gibson", newLocation, Collections.singletonList(OWNER_1));
-		final Brand actual = smof.find(Brand.class).byElement(brand);
-		assertEquals(expected, actual);
-	}
-
-	@Test
-	public void testUpdateSetSameFieldTwice() {
-		final Brand brand = Brand.create("Gibson", new Location("Nashville", "USA"), Collections.singletonList(OWNER_1));
-		smof.insert(brand);
-		Location newLocation1 = new Location("New York", "USA");
-		Location newLocation2 = new Location("Los-Angeles", "USA");
-		smof.update(Brand.class)
-		.set(Brand.LOCATION, newLocation1)
-		.set(Brand.LOCATION, newLocation2)
-		.where()
-		.fieldEq(Brand.NAME, "Gibson")
-		.execute();
-		final Brand expected = Brand.create("Gibson", newLocation2, Collections.singletonList(OWNER_1));
-		final Brand actual = smof.find(Brand.class).byElement(brand);
-		assertEquals(expected, actual);
-	}
-
 	@Test
 	public void testSmofGridRef() throws IOException {
 		final SmofGridStreamManager gridStream = smof.getGridStreamManager();
